==> supplemental_materials/scripts/04-Migration_Observations_Report.Rmd <==
---
title: Juvenile Salmon Migration Observations in the Discovery Islands in 2020
output: word_document
---

```{r setup, include = FALSE, messages = FALSE}
knitr::opts_chunk$set(fig.pos = 'H', echo = FALSE, warning = FALSE, message = FALSE, error = FALSE)

library(tidyverse)
library(lubridate)
library(knitr)
library(here)
library(car)
library(ggridges)
library(hakaiR) 
# install hakaiR from github: 
#install.packages("devtools")
#devtools::install_github("HakaiInstitute/hakaiR")


     # Survey Data
survey_seines <- read_csv(here("supplemental_materials", "report_data", "survey_seines.csv")) %>%
  rename("Sockeye" = "so_total", "Pink" = "pi_total", "Chum" = "cu_total")
lice_bootstraps <- read_csv(here("supplemental_materials", "report_data", "lice_bootstraps.csv"))
# Temp processed_data
qu39_ctd <- read_csv(here("supplemental_materials", "report_data", "qu39_ctd.csv"))
temperature_anomaly_data <- read_csv(here("supplemental_materials", "report_data", "temperature_anomaly_data.csv"))
min_max_data <- read_csv(here("supplemental_materials", "report_data", "min_max_temps.csv"))
average_temps <- read_csv(here("supplemental_materials", "report_data", "average_temps.csv"))
proportion <- read_csv(here("supplemental_materials", "report_data", "proportion.csv"))
# Migration Timing processed_data
predict_average_prop <- read_csv(here("supplemental_materials", "report_data", "predict_average_prop.csv"))
predict_annual_prop <- readRDS(here("supplemental_materials", "report_data", "predict_annual_prop.RDS"))
peak_dates <- read_csv(here("supplemental_materials", "report_data", "peak_dates.csv"))
catch_intensity <- read_csv(here("supplemental_materials", "report_data", "catch_intensity.csv"))
cu_cum_abund_annual_DI_2017 <- read_csv(here("supplemental_materials", "report_data", "cu_cum_abund_DI_2017.csv"))
# Length processed_data
length_histo <- read_csv(here("supplemental_materials", "report_data", "length_histo.csv"))
fish_cond <- read_csv(here("supplemental_materials", "report_data", "fish_cond.csv"))
fish_cond$species <- factor(fish_cond$species, levels = c("SO", "PI", "CU", "CO"))
# Must match the order of the species re ordered above
levels(fish_cond$species) <- c('Sockeye', 'Pink', 'Chum', 'Coho')

# Standardized z-scores
heatmap_data <- readRDS(here("supplemental_materials", "report_data", "heatmap_data.RDS"))


spp_labels <- c(CU = "Chum", PI = "Pink", SO = "Sockeye", DI = "Discovery Islands",
                JS = "Johnstone Strait")

pl_colours <- c("pi_total" = "pink", 
                "so_total" = "#00BFC4",
                "cu_total" = "#7CAE00",
                "co_total" = "#F8766D",
                "he_total" = "#C77CFF")

current_year <- max(survey_seines$year, na.rm = T)
project_years <- current_year - 2015 + 1
study_range <- paste(2015, "-", current_year)

```

```{r migration timing stats, include=FALSE}

# calcualte the time series average quartiles for each species
tsa_migration_timing <- peak_dates %>% 
  group_by(species) %>% 
  summarise(q1= mean(q1),
            median= mean(median),
            q3= mean(q3),
            n = n()) %>% 
  mutate(year = paste(2015, "-", current_year), region = "DI") %>% 
  ungroup()%>% 
  arrange(year, desc(species))

# combine time series average with annual observations into one df
migration_timing <- peak_dates %>% 
  mutate_if(is.numeric, round) 

# Define study period and current year peak migration date variables for inline code text in report. 

# Sockeye
so_q2_DI_TSA <- as.numeric(migration_timing %>% 
  filter(species == "SO" & year == paste(2015, "-", current_year) & region == "DI") %>% 
  select(median)) 

so_q2_DI_current <- as.numeric(migration_timing %>% 
  filter(species == "SO" & year == current_year & region == "DI") %>% 
  select(median))

# Pink
pi_q2_DI_TSA <- as.numeric(migration_timing %>% 
  filter(species == "PI" & year == paste(2015, "-", current_year) & region == "DI") %>% 
  select(median))

pi_q2_DI_current <- as.numeric(migration_timing %>% 
  filter(species == "PI" & year == current_year & region == "DI") %>% 
  select(median))

# Chum
cu_q2_DI_TSA <- as.numeric(migration_timing %>% 
  filter(species == "CU" & year == paste(2015, "-", current_year) & region == "DI") %>% 
  select(median))

cu_q2_DI_current <- as.numeric(migration_timing %>% 
  filter(species == "CU" & year == current_year & region == "DI") %>% 
  select(median))

# calculate the differene in timing for current year species in DI compated to study period
so_timing_diff <- abs(so_q2_DI_current - so_q2_DI_TSA)

# Define whether the timing was earlier or later for use in inline text
so_later_or_earlier <- ifelse(so_q2_DI_current - so_q2_DI_TSA > 0, "later", "earlier")

# calculate the differene in timing for current year piecies in DI compated to study period
pi_timing_diff <- abs(pi_q2_DI_current - pi_q2_DI_TSA)

# Define whether the timing was earlier or later for use in inline text
pi_later_or_earlier <- ifelse(pi_q2_DI_current - pi_q2_DI_TSA > 0, "later", "earlier")

# calculate the differene in timing for current year cuecies in DI compated to study period
cu_timing_diff <- abs(cu_q2_DI_current - cu_q2_DI_TSA)

# Define whether the timing was earlier or later for use in inline text
cu_later_or_earlier <- ifelse(cu_q2_DI_current - cu_q2_DI_TSA > 0, "later", "earlier")

# Define function f here to convert Julian Days to dates
f <- function(x) {
  format(x + as.Date("2018-01-01") -1, format = "%B %d")
}

migration_timing <- peak_dates %>% 
  rename(Year = year, Species = species, Region = region, n = n, Q1 = q1, "Peak Date" = median,
         Q3 = q3) %>% 
  select(Year, Region, Species, Q1, "Peak Date", Q3, Spread) %>% 
  mutate_if(is.numeric, round) %>% 
  mutate(Q1 = f(Q1),
         "Peak Date" = f(`Peak Date`),
         Q3 = f(Q3))
         

```

```{r catch intensity}
current_catch_intensity <- catch_intensity %>% 
  group_by(species) %>% 
  mutate(rank =  row_number(-mean_catch)) %>% 
  ungroup() %>% 
  filter(year == current_year) %>% 
  mutate(species = replace(species, species == "Sockeye", "sockeye")) %>% 
  mutate(species = replace(species, species == "Pink", "pink")) %>% 
  mutate(species = replace(species, species == "Chum", "chum")) %>% 
  arrange(desc(mean_catch))

cu_ci <- current_catch_intensity %>%
  filter(species == "chum") %>% 
  select(mean_catch) %>% 
  as.numeric()

pi_ci <- current_catch_intensity %>% 
  filter(species == "pink") %>% 
  select(mean_catch) %>% 
  as.numeric()

so_ci <- current_catch_intensity  %>% 
  filter(species == "sockeye") %>% 
  select(mean_catch) %>% 
  as.numeric()
```

```{r species proportions stats}

current_prop <- proportion %>% 
  filter(year == current_year) %>% 
  filter(species %in% c("so_total", "pi_total", "cu_total")) %>% 
  mutate(species = replace(species, species == "so_total", "sockeye")) %>% 
  mutate(species = replace(species, species == "pi_total", "pink")) %>% 
  mutate(species = replace(species, species == "cu_total", "chum")) %>% 
  arrange(desc(proportion))

 
new_spp_names <- dplyr::recode(proportion$species, so_total = "Sockeye",
                               pi_total = "Pink", cu_total = "Chum",
                               he_total = "Herring", co_total = "Coho",
                               ck_total = "Chinook")

proportions_table <- proportion %>% 
  mutate(species = new_spp_names,
         proportion = round(proportion, 3)) %>% 
  rename(Year = year, Species = species, Proportion = proportion) %>% 
  select(Year, Species, Proportion) %>% 
  spread(key = Species, value = Proportion)

```

```{r sea lice stats}

lice_bootstraps$species <- factor(lice_bootstraps$species) %>% 
  fct_relevel("SO", "PI", "CU") %>% 
  fct_recode("Sockeye" = "SO", "Pink" = "PI", "Chum" = "CU")

df <- lice_bootstraps %>% 
  group_by(year, region, species, louse_species) %>% 
  summarize(mean_abund = mean(estimate),
            upper_ci = mean(`97.5%`),
            lower_ci = mean(`2.5%`),
            n = n()) %>% 
  group_by(region, species, louse_species) %>% 
  mutate(rank =  row_number(-mean_abund)) %>% 
  ungroup()

pretty_sealice_df <- df %>% 
  ungroup() %>% 
  mutate(
         year = factor(year),
         lower_ci = round(lower_ci, 2),
         upper_ci = round(upper_ci, 2),
         "95% CI" = paste(lower_ci, "-", upper_ci)) %>% 
  select("Year" = "year", "Region" = "region", "Species" = "species",
         "Louse Species" = "louse_species", "Mean Abundance" = "mean_abund",
         "95% CI") %>% 
  mutate_if(is.numeric, round, 2)

current_sealice_stats <- pretty_sealice_df %>% 
  filter(Year == current_year) %>% 
  mutate(abund_ci = paste(`Mean Abundance`, "(95% CI", `95% CI`, ")"))

so_cal <- current_sealice_stats %>% 
  filter(Species == "Sockeye",
         `Louse Species` == "caligus") %>% 
  select(`Mean Abundance`) %>% 
  as.numeric()

so_lep <- current_sealice_stats %>% 
  filter(Species == "Sockeye",
         `Louse Species` == "lep") %>% 
  select(`Mean Abundance`) %>% 
  as.numeric()

pi_cal <- current_sealice_stats %>% 
  filter(Species == "Pink",
         `Louse Species` == "caligus") %>% 
  select(`Mean Abundance`) %>% 
  as.numeric()

pi_lep <- current_sealice_stats %>% 
  filter(Species == "Pink",
         `Louse Species` == "lep") %>% 
  select(`Mean Abundance`) %>% 
  as.numeric()

cu_cal <- current_sealice_stats %>% 
  filter(Species == "Chum",
         `Louse Species` == "caligus") %>% 
  select(`Mean Abundance`) %>% 
  as.numeric()

cu_lep <- current_sealice_stats %>% 
  filter(Species == "Chum",
         `Louse Species` == "lep") %>% 
  select(`Mean Abundance`) %>% 
  as.numeric()
```

```{r length stats}
length_histo_current <- length_histo %>% 
  filter(year == current_year,
         region == "DI") %>% 
  mutate(category = current_year)

len_tt <- length_histo %>% 
  filter(region == "DI",
         year  != current_year) %>% 
  mutate(category = 'tsa') %>% 
  rbind(length_histo_current) %>% 
  group_by(species) %>% 
  do(broom::tidy(t.test(.$fork_length ~ .$category))) %>% 
  arrange(desc(species))

summary_lengths <- length_histo %>% 
  filter(region == "DI") %>% 
  group_by(year, region, species) %>% 
  summarize(SD = sd(fork_length, na.rm = TRUE),
            fork_length = round(mean(fork_length),1),
            n = n()) %>% 
  mutate(SE = SD / sqrt(n),
         CI = qt(1 - (0.05 / 2), n - 1) * SE) %>% 
  arrange(year, desc(species)) %>% 
  select(year, species, n, fork_length, CI) %>% 
  mutate_if(is.numeric, round, 1) %>% 
  ungroup() %>% 
  mutate(year = as.character(year))
  


```

```{r sst stats}
# SST z-score
## I'm going to filter the time period down to May and Juen because that's the most relevant period that juvenile salmon are likely to be in the Strait of Georgia

sst_annual_mean <- qu39_ctd %>%
  ungroup() %>%
  mutate(date = as_date(date)) %>%
  mutate(month = month(date), year = year(date)) %>%
  filter(month >= 5 & month <= 6, station == "QU39") %>%
  select(year,  yday, mean_temp) %>%
  group_by(year) %>%
  summarise(sd_temp = sd(mean_temp),
    mean_temp = mean(mean_temp, na.rm = T),
    n = n(),
    yday  = mean(yday))
            

sst_mean <- mean(sst_annual_mean$mean_temp)
sst_sd <- sd(sst_annual_mean$mean_temp)

sst_z <- (sst_annual_mean$mean_temp - sst_mean) / sst_sd

sst_time_series_stats <- qu39_ctd %>%
  ungroup() %>%
  mutate(date = as_date(date)) %>%
  mutate(month = month(date), year = year(date)) %>%
  filter(month >= 5 & month <= 6, station == "QU39") %>%
  select(year, mean_temp) 

sst_current_year <- sst_time_series_stats %>% 
  filter(year == current_year)

sst_stats_table <- broom::tidy(t.test(sst_current_year$mean_temp, sst_time_series_stats$mean_temp))

sst_nice_table <- sst_annual_mean %>% 
  mutate(mean_temp = round(mean_temp, 2),
         sd_temp = round(sd_temp, 2)) %>% 
  select(Year = year, "Temperature (C)" = mean_temp, SD = sd_temp) 

```

```{r map, fig.cap = "Sampling locations in 2019", out.width="90%"}
include_graphics(here('supplemental_materials', 'map.png'))
```

# Abstract 

The Hakai Institute Juvenile Salmon Program has been monitoring juvenile salmon migrations in the Discovery Islands and Johnstone Strait in British Columbia, Canada since 2015 with the specific purpose to understand how ocean conditions experienced by juvenile salmon during their early marine migration impact their growth, health, and ultimately survival. This report summarizes migration timing, purse seine catch intensity and composition, fish length and weight, sea-louse loads, and ocean temperatures observed from `r project_years` years of this research and monitoring program. Migration timing for sockeye, pink, and chum was not significantly different than respective time-series averages and occurred on `r f(so_q2_DI_current)` for sockeye, `r f(cu_q2_DI_current)` for chum, and on `r f(pi_q2_DI_current)` for pink salmon. In order of highest to smallest catch proportion seines were dominated by juvenile `r current_prop[[1,2]]`, `r current_prop[[2,2]]`, and `r current_prop[[3,2]]` salmon. Catch intensity (our relative abundance measurement) for chum was the lowest on record in the time series, though sockeye and chum catch intensity was nearer their time series averages. Mean annual fork length for sockeye, pink, and chum salmon were all shorter than the time series average. The abundance of pre-adult and adult _Caligus clemensi_ sea lice was relatively high on juvenile sockeye, pink, and chum salmon in 2020 compared to previous years. The salmonid-specialist sea louse, _Lepeophtheirus salmonis_, had relatively high abundance on pink and chum salmon but low abundance on sockeye. May--June 30 m depth integrated ocean temperature in the northern Strait of Georgia in `r current_year` was `r round(sst_stats_table$estimate, 2)` °C `r ifelse(sst_stats_table$estimate > 0, "warmer",  "colder")` than average for the time series (`r study_range`), the warmest we've observed in our time-series.

# Results

## Migration Timing
Sockeye:  `r f(so_q2_DI_current)`, `r round(so_timing_diff)` days `r so_later_or_earlier` than the time series average of `r f(so_q2_DI_TSA)`.
Pink:  `r f(pi_q2_DI_current)`, `r round(pi_timing_diff)` days `r pi_later_or_earlier` than the average of `r f(pi_q2_DI_TSA)`. 
Chum: `r f(cu_q2_DI_current)`, `r round(cu_timing_diff)` day `r cu_later_or_earlier` than the average of `r f(cu_q2_DI_TSA)`. 


## Catch intensity 

Sockeye: `r so_ci` 

Pink: `r  pi_ci` 

chum: `r cu_ci` 

## Catch proportion

Catch proportion was dominated by `r current_prop[[1,2]]` with `r round(current_prop[[1,4]] * 100, 1)`% of the catch.

Second highest catch proportion was `r current_prop[[2,2]]` salmon at `r round(current_prop[[2,4]] * 100, 1)`%.

Third highest was `r current_prop[[3,2]]` salmon at `r round(current_prop[[3,4]] * 100, 1)`% 

## Lengths

Sockeye length in `r current_year` was `r round(len_tt[1,3],1)` mm which compared to the time series average of `r round(len_tt[1,4],1)` mm results in a `r ifelse(len_tt$p.value[1] <  0.05, paste("statistically significant"), paste("statistically insignificant"))` difference of `r round(len_tt[1,2],1)`   mm (_p_ = `r round(len_tt$p.value[1],2)`, 95% CI `r round(len_tt$conf.low[1],1)`--`r round(len_tt$conf.high[1],1)` mm). 

Average pink lengths in `r current_year` was `r round(len_tt[2,3],1)` mm which compared to the time series average of `r round(len_tt[2,4],1)` mm results in a `r ifelse(len_tt$p.value[2] <  0.05, paste("statistically significant"), paste("statistically insignificant"))` difference of `r round(len_tt[2,2],1)`   mm (_p_ = `r round(len_tt$p.value[2],2)`, 95% CI `r round(len_tt$conf.low[2],1)`--`r round(len_tt$conf.high[2],1)` mm). 

Average chum lengths in `r current_year` was `r round(len_tt[3,3],1)` mm which compared to the time series average of `r round(len_tt[3,4],1)` mm results in a `r ifelse(len_tt$p.value[3] <  0.05, paste("statistically significant"), paste("statistically insignificant"))` difference of `r round(len_tt[3,2],1)`   mm (_p_ = `r round(len_tt$p.value[3],2)`, 95% CI `r round(len_tt$conf.low[3],1)`--`r round(len_tt$conf.high[3],1)` mm). 

## Sea lice

Sockeye salmnoh had an average of `r so_cal` _C. clemensi_ per fish and `r so_lep` _L. salmonis_ per fish.

Chum salmon had `r cu_cal`  _Caligus clemensi_ per fish and `r cu_lep` _Lepeophtheirus salmonis_ per fish.  

Pink salmon had `r pi_cal`  _C. clemensi_ per fish, and `r  pi_lep`  _L. salmonis_ per fish. 

## Ocean temps

Ocean temperature in the top 30 m of the water column in May and June during the juvenile salmon out-migration at QU39 in the northern Strait of Georgia was  `r round(sst_stats_table[1,2],1)` °C, `r round(sst_stats_table$estimate, 2)` `r ifelse(sst_stats_table$estimate > 0,  paste("warmer"), paste("colder"))` °C than average.


```{r heatmap,  fig.cap = paste("The number of standard deviations (z score) from the time series average (2015-", current_year, " for key migration parameters. Size and colour saturation of circles indicates the magnitude of the anomaly. Blue colour indicates less than average; grey indicates average; red indicates greater than average. Peak migration date is based on the median date of fish capture in the Discovery Islands. Length is based on the average fork length from the Discovery Islands and Johnstone Strait combined. Parasite load is the average abundance of all sea-louse species in their motile life stages for both the Discovery Islands and Johnstone Strait regions. Ocean temperature describes the mean ocean temperature in the top 30 m at station QU39 in the northern Strait of Georgia in May and June.", sep=""), out.width = "100%", fig.height=11, fig.width=10}

cols <- c("#4575b4",
          "lightgrey",
          "#d73027"
          )

heatmap_data$measure <- heatmap_data$measure %>%
    fct_recode("Fork Lengths" = "Length", "Sea-louse Abundance" = "Sea lice Abundance", "Ocean Temperature" = "Sea-surface Temperature")  %>% 
  fct_relevel("Migration Timing", "Catch Intensity", "Fork Lengths", "Sea-louse Abundance", "Ocean Temperature")
#TODO: Enhancement - make it so the legend show 5 points of size
(ggplot(heatmap_data, aes(year, spp, colour = Z)) +
  geom_point(size = abs(10 * heatmap_data$Z), alpha = 0.00001) +
  geom_point(aes(size = abs(10 * Z)), show.legend = F) +
  scale_size_continuous(range = c(3,15)) +
  ylab("") +
  xlab("Year") +
  scale_colour_gradientn(colours = cols) +
  guides(color = guide_legend(override.aes = list(size = c(10, 3, 5, 10), alpha = 1)), name = "Z score") +
  labs(color = "Z Score") +
  scale_y_discrete(expand = expand_scale(mult = c(.75, .75))) +
  scale_x_continuous(expand = expand_scale(mult = c(.1, .1)), n.breaks = project_years) +
  #theme(panel.grid.major.y= element_line("grey")) +
  facet_wrap(. ~ measure, ncol = 1, scales = "free") +
  hakademic_theme(base_size = 18)+
  theme(legend.position = "bottom")
)

ggsave(here("supplemental_materials", "figs", "heatmap.png"), width = 13, height = 11)
```


```{r migration-timing-plot, fig.cap = paste("Cumulative catch of sockeye, pink, and chum, in the Discovery Islands and Johnstone Strait between 2015 and ", current_year, ".", sep=""), out.width="100%"}
shapes <-
  c(
    "2015 - 2020" = "triangle",
    "2021" = "circle",
    "2020" = "circle",
    "2019" = "circle",
    "2018" = "circle",
    "2017" = "circle",
    "2016" = "circle",
    "2015" = "circle"
  )

predict_average_prop <- predict_average_prop %>% 
  filter(region == "DI")

predict_average_prop$species <- factor(predict_average_prop$species) %>% 
  fct_relevel(c("Sockeye", "Pink", "Chum")) # contains only DI

predict_annual_prop$species <- factor(predict_annual_prop$species) %>% 
  fct_relevel(c("Sockeye", "Pink", "Chum"))
predict_annual_prop$year <- factor(predict_annual_prop$year)

cu_cum_abund_annual_DI_2017$species <- factor(cu_cum_abund_annual_DI_2017$species)

# Remove 2017 chum logistic model, in favour of loess model.
predict_annual_prop <- predict_annual_prop %>% 
  filter(year == current_year)

ggplot()+
      labs(x = 'Date', y = 'Cumulative Catch %') +
  geom_line(data = predict_average_prop, aes(x = x, y = y, color = year),
  size = 1.25, alpha = 0.75) +
  geom_line(data = predict_annual_prop, aes(x = x, y = y,
                                                color = year),
                linetype= 1, size = 1.25, alpha = 1) +
  scale_x_continuous(breaks = c(135, 152, 166, 182, 196), 
                     labels = c("May 15", "June 1", "June 15", 
                                "July 1", "July 15")) +
      coord_cartesian(xlim = c(128, 190)) + 
  scale_color_manual(values = c("grey", "black")) +
  theme(legend.position = c(0.9, 0.5)) +
  facet_grid(species ~ .)
    
ggsave(here("supplemental_materials", "figs", "migration_timing.png"), width = 9, height = 6)
```

```{r catch-intensity-plot, fig.cap = "The catch intensity (our proxy for abundance) of sockeye, pink, and chum salmon in the Discovery Islands. Numbers under each bar indicate the number of seines in which the species was caught, and error bars indicate the 95 percent confidence region.", out.width="90%"}

catch_intensity$species <- catch_intensity$species %>% 
  fct_relevel("Sockeye", "Pink", "Chum")

(ggplot(catch_intensity, aes(x = factor(year), y = mean_catch, fill = species)) +
  geom_bar(colour = "black",  stat="identity", position=position_dodge()) +
  geom_errorbar(aes(ymin = mean_catch - se, ymax = mean_catch + se),
                    width = 0.2,
                    position = position_dodge(0.9)) +
  geom_text(aes(y = 0.0, label = paste0(n)), size = 3.15, vjust = 1.25,
                position = position_dodge(0.9)) +
  xlab("Year") +
  ylab("Catch Intensity") +
  labs(fill = "Species") +
  scale_fill_manual(values = hakaiR::hakai_palette("hakai_salmon")))


ggsave(here("supplemental_materials", "figs", "catch_intesity.png"))
```

```{r proportion-plot, out.width="90%", fig.cap="The annual proportion of fish captured in the Discovery Islands and Johnstone Strait combined."}
    
proportion$species <- proportion$species %>% 
  fct_relevel("so_total", "pi_total", "cu_total", "co_total", "he_total") %>% 
  fct_recode("Sockeye" = "so_total", "Pink" = "pi_total", "Chum" = "cu_total", "Coho" = "co_total",
             "Herring" = "he_total")

    ggplot(data = proportion, aes(x = year, y = proportion, fill = species)) +
      geom_bar(colour = "black", stat="identity", position = 'stack') +
      xlab("Year") +
      ylab("Proportion") +
      scale_fill_manual(values = hakai_palette("hakai_salmon")) +
      labs(fill = "Species") +
      scale_x_continuous(n.breaks = project_years)
      

ggsave(here("supplemental_materials", "figs", "proportions.png"))
```

```{r length-plot, out.width="90%", fig.cap="Distributions of juvenile salmon fork lengths for each year in the Discovery Islands and Johnstone Strait. Note that these distributions contain multiple age classes."}

length_histo$species <- length_histo$species %>% 
  fct_relevel("SO", "PI", "CU", "CO")

DI_length_histo <- length_histo %>% 
  filter(region == "DI")

ggplot(DI_length_histo, aes(fork_length, y = fct_rev(factor(year)), fill = species)) +
  geom_density_ridges(color='black', scale = 2.5, alpha = 0.9) +
  xlab("Fork Length (mm)") +
  facet_grid(species ~ ., labeller = labeller(region = spp_labels, species = spp_labels)) +
  ylab("Year") +
  #scale_fill_hakai() +
  scale_y_discrete(expand = expand_scale(add = c(0.2, 2.8))) +
  coord_cartesian(xlim = c(60, 160)) +
  guides(fill = FALSE) +
  scale_fill_manual(values = hakai_palette("hakai_salmon"))

ggsave(here("supplemental_materials", "figs", "lengths.png"), width = 8, height = 7)
```


```{r condition-plot, fig.cap=paste("Length and weight regressions for juvenile salmon caught in the Discovery Islands and Johnstone Strait in", current_year, "coloured red, compared to all other years in black.")}
library(cowplot)
library(ggpubr)
time_series_fish_cond <- fish_cond %>% 
  mutate(year = factor(year(survey_date))) %>% 
  filter(species %in% c("Sockeye", "Pink", "Chum", "Coho"),
         year != current_year) %>% 
  mutate(Years = "Time Series")

current_fish_cond <- fish_cond %>% 
  filter(year == current_year) %>% 
  mutate(Years = factor(current_year))


#TODO: Enhancement - Add histograms to x and y axes

ggplot(data = fish_cond, aes(x = fork_length, y = weight, colour= Years)) +
  geom_point(data = time_series_fish_cond, alpha = .2, size = 3) +
  geom_point(data = current_fish_cond, alpha = .3, size = 2) + 
  ylab('Weight (g)')+
  xlab('Fork Length (mm)')+
  coord_cartesian(xlim = c(50, 177), ylim = c(0, 80)) +
  scale_color_manual(values=c("#B52025", "black")) +
  #theme(legend.position = c(.9, .9), legend.title=element_blank()) +
  facet_wrap(~species, nrow=2)

ggsave(here("supplemental_materials", "figs", "length_weight.png"), width = 8, height = 6)

```


```{r sealice-abundance-plot, fig.cap="The abundance of motile sea lice on juvenile salmon in the Discovery Islands and Johnstone Strait. The numbers under each bar indicate the sample size and the error bars indicate the 95 percent confidence region.", out.width="85%"}

df %>% 
  filter(region != "JS") %>% 
  mutate(louse_species = replace(louse_species, louse_species == "caligus", "C. clemensi")) %>% 
  mutate(louse_species = replace(louse_species, louse_species == "lep", "L. salmonis")) %>% 
ggplot(aes(x = year, y = mean_abund, fill = louse_species)) +
      geom_bar(colour = "black", stat = "identity", position = position_dodge()) +
      geom_errorbar(aes(ymin = mean_abund - ((upper_ci - lower_ci) / 2), 
                        ymax = mean_abund + ((upper_ci - lower_ci) / 2)), 
                    width = 0.2,
                    position = position_dodge(0.9)) +
      facet_grid(. ~ species,
                 labeller = labeller(region = spp_labels)) +
      labs(x = "Year", y = "Abundance") +
      scale_fill_manual(values = hakai_palette("hakai_web")) +
      theme(legend.position="bottom") +
      # geom_text(aes(label = ifelse(Abundance == 0, round(Abundance, 1), '')),
      #          position = position_dodge(0.9), vjust = -0.5) +
      # geom_text(aes(y = 0.0, label = paste0(n)), size = 3.15, vjust = 1.25,
      #           position = position_dodge(0.9)) +
      #scale_y_continuous(expand = expand_scale(mult = c(.15, 0))) +
      #scale_fill_hakai() +
      labs(fill = "Louse Species") + 
  theme(legend.text = element_text(face = "italic", size = 10))+
  scale_x_continuous(n.breaks = project_years)+
  theme(axis.text.x = element_text(angle=90, vjust=.5, hjust=1))

  #coord_cartesian(ylim = c(0, 1.7)) +


ggsave(here("supplemental_materials", "figs", "sealice_abundance.png"), width = 7, height = 4)
```


```{r sst-plot, out.width="90%", fig.cap=paste("Thirty-meter depth-integrated ocean temperatures at station QU39 in the northern Strait of Georgia is the solid black line which represents average temperatures from 2015–2019. Blue areas represent temperatures from ", current_year, "that are below the 2015–", current_year, "average and red areas represent above average temperatures. The shaded grey area is 1 SE of the loess regression. The black dots are the daily minimum and maximum temperatures observed over the time series.", sep="")}


ggplot(data = temperature_anomaly_data, aes(x = yday, y = mean_temp)) +
      geom_point(aes(x = yday, y = predicted_mean_temp), size = 0.1)+
      geom_line(aes(x = yday, y = predicted_mean_temp), size = 1 ) +
      geom_ribbon(data = subset(temperature_anomaly_data, 
                                mean_temp >= predicted_mean_temp),
                  aes(ymin = predicted_mean_temp, ymax = mean_temp),
                  fill = "#d73027", size = 1)+
      geom_ribbon(data = subset(temperature_anomaly_data, 
                                mean_temp <= predicted_mean_temp), 
                  aes(ymin = mean_temp, ymax = predicted_mean_temp),
                  fill = "#4575b4", size = 1)+
      # geom_smooth(data = average_temps, aes(x = yday, y = mean_temp),
      #             size = 1, colour = 'black', se = T, span = .65) +
      # geom_point(data = min_max_data,
      #            aes(x = yday, y = min_temp), size = 0.5) +
      scale_x_continuous(breaks = (c(32, 60, 91, 121, 152, 182, 213)),
                         labels = (c("Feb", "Mar", "Apr", "May", "Jun", 
                                     "Jul", "Aug"))) +
      labs(x = "Date", y = "Temperature [°C]") +
      coord_cartesian(xlim = c(32,213))

ggsave(here("supplemental_materials", "figs", "sst.png"), width = 8, height = 6)

```

# Data

Some of the models and methods used to produce statistics and visualizations in this report will undergo ongoing development as we seek to improve the accuracy of current estimates, incorporate historical observations from other organizations, and add new variables to our annual observations. The development of our annual observations data, code, and analyses can be observed and contributed to from our code and data repository at https://github.com/HakaiInstitute/jsp-time-series. Summarized data can be observed in the tables below or accessed digitally at the web-address mentioned above and utilized under the Creative Commons Attribution 4.0 International License (CC-BY 4.0). The R code written to produce the figures and statistics in this report can be accessed in our repository and used to reproduce, contribute to, or alter this analysis. An interactive data explorer tool for this developing time series can be accessed online at http://hecate.hakai.org/shiny/JSP/.

# Tables


```{r migration-timing-table}
migration_timing$Species <- factor(migration_timing$Species) %>% 
  fct_relevel("SO", "PI", "CU") %>% 
  fct_recode("Sockeye" = "SO", "Pink" = "PI", "Chum" = "CU")

migration_timing <- migration_timing %>% 
  drop_na() %>% 
  filter(Year %in% c(paste("2015", "-", current_year), current_year))
knitr::kable(migration_timing, format = 'markdown')
```
Table: (\#tab:migration-timing-table) Migration timing statistics for the cumulative catch of sockeye, pink, and chum salmon in the Discovery Islands in 2019, compared to the time-series average (2015 - `r current_year`). Q1 is when 25 % of the species passed through the regions, peak date is the median when 50 % passed through, Q3 is 75%, and Spread is the difference between Peak Date and Q1. The region DI indicates the Discovery Islands while for species SO is sockeye, PI is pink, and CU is chum. 

```{r length-table}
summary_lengths$species <- factor(summary_lengths$species) %>% 
  fct_relevel("SO", "PI", "CU") %>% 
  fct_recode("Sockeye" = "SO", "Pink" = "PI", "Chum" = "CU")
summary_lengths <- summary_lengths %>% 
  select(Year = year, Species = species, N = n, "Fork Length" = fork_length, CI = CI)
knitr::kable(summary_lengths, format = 'markdown')
```
Table: (\#tab:length-table) Mean fork lengths for each year, species, and region with the +/- 95 % confidence interval (95% CI). The column n indicates the number of fish measured.


```{r sst table}
kable(sst_nice_table, fomrat = 'markdown')
```

# References
